% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_unzip.R
\name{prism_unzip}
\alias{prism_unzip}
\title{Unzips a .prism file}
\usage{
prism_unzip(prism_path, output_dir = dirname(prism_path))
}
\arguments{
\item{prism_path}{Path to the `.prism` file to extract.}

\item{output_dir}{Optional parent directory in which to create the `_unzipped`
folder. Defaults to the directory containing `prism_path`.}
}
\value{
Invisibly returns the full path to the folder where the files were unzipped.
}
\description{
Given a path to a “.prism” file, this function will create a sibling “_unzipped” folder (named after the file) and unzip everything into it.
}
