% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzfx_to_csv.R
\name{pzfx_to_csv}
\alias{pzfx_to_csv}
\title{Convert .pzfx Files to CSV}
\usage{
pzfx_to_csv(file_path, output_dir = getwd())
}
\arguments{
\item{file_path}{A character string specifying the path to the .pzfx file to be processed.
It must point to a valid file on your system.}

\item{output_dir}{A character string specifying the directory where the CSV files should be saved.
Defaults to the current working directory (`getwd()`).}
}
\value{
None. The function saves the extracted data as CSV files in the specified directory.
}
\description{
This function reads tables from a GraphPad Prism (.pzfx) file, cleans the data by removing rows with all NA values,
and saves each table as a separate CSV file. It provides a simple way to convert the data from GraphPad Prism
files into a more commonly used format like CSV for further analysis.
}
\details{
The function first lists all the tables in the given .pzfx file, extracts the data from each table,
and removes rows with missing values using the `clean_na_rows` function. The cleaned data for each table
is then saved as a separate CSV file.
}
\examples{
# Example usage:
\dontrun{
pzfx_to_csv("path/to/your/file.pzfx", "path/to/save/csvs")}
}
