% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzfx_to_xlsx.R
\name{pzfx_to_xlsx}
\alias{pzfx_to_xlsx}
\title{Convert .pzfx Files to XLSX}
\usage{
pzfx_to_xlsx(file_path, output_dir = getwd())
}
\arguments{
\item{file_path}{A character string specifying the path to the .pzfx file to be processed.
It must point to a valid file on your system.}

\item{output_dir}{A character string specifying the directory where the Excel file should be saved.
Defaults to the current working directory (`getwd()`).}
}
\value{
None. The function saves the extracted data as an Excel file in the specified directory.
}
\description{
This function reads tables from a GraphPad Prism (.pzfx) file, cleans the data by removing rows with all NA values,
and saves each table as a separate sheet in an Excel workbook (.xlsx format). It provides a simple way to convert the
data from GraphPad Prism files into a more commonly used format like Excel for further analysis.
}
\details{
The function first lists all the tables in the given .pzfx file, extracts the data from each table,
and removes rows with missing values using the `clean_na_rows` function. Each table is then added as a separate
sheet in an Excel workbook. The workbook is then saved in the specified directory.
}
\examples{
# Example usage:
\dontrun{
pzfx_to_xlsx("path/to/your/file.pzfx", "path/to/save/excel")}

}
